+++ src/views/index.js
@@ -46,6 +46,21 @@ exports.authorView = ({
   const mention = `[@${name}](${feedId})`
   const markdownMention = highlightJs.highlight('markdown', mention).value
 
+  const areFollowing = relationship === 'you are following'
+
+  const contactFormType = areFollowing
+    ? 'unfollow'
+    : 'follow'
+
+  // We're on our own profile!
+  const contactForm = relationship !== null
+    ? form({ action: `/${contactFormType}/${encodeURIComponent(feedId)}`, method: 'post' },
+      button({
+        type: 'submit'
+      },
+      contactFormType))
+    : null
+
   const prefix = section({ class: 'message' },
     header({ class: 'profile' },
       img({ class: 'avatar', src: avatarUrl }),
@@ -59,7 +74,8 @@ exports.authorView = ({
       : null,
     footer(
       a({ href: `/likes/${encodeURIComponent(feedId)}` }, 'view likes'),
-      span(relationship)
+      span(relationship),
+      contactForm
     )
   )